requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
opencv-python
numpy
torch
git+https://github.com/openai/CLIP.git